server/adisyon_project/requirements.txt
anyascii==0.3.2
asgiref==3.8.1
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.0
backcall==0.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.2
contourpy==1.1.0
cryptography==41.0.1
cssselect2==0.7.0
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==2.0.0
Django==5.2
django-allauth==0.54.0
django-classy-tags==3.0.1
django-cms==3.8.2
django-extensions==4.1
django-filer==3.0.3
django-filter==22.1
django-formtools==2.4
django-friendship==1.9.6
django-js-asset==2.0.0
django-modelcluster==6.0
django-mptt==0.14.0
django-permissionedforms==0.1
django-polymorphic==3.1.0
django-rest-auth==0.9.5
django-sekizai==3.0.1
django-taggit==3.1.0
django-treebeard==4.4
djangocms-admin-style==2.0.2
djangocms-attributes-field==3.0.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.1
djangocms-googlemap==2.1.0
djangocms-icon==2.0.0
djangocms-installer==2.0.0
djangocms-link==3.1.1
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
draftjs-exporter==2.1.7
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.8.0
filetype==1.2.0
firebase-admin==6.1.0
fonttools==4.42.0
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.14.0
html5lib==1.1
httplib2==0.22.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.20.2
ipython==8.9.0
jedi==0.18.2
jupyter_client==8.0.1
jupyter_core==5.1.5
kiwisolver==1.4.4
l18n==2021.3
lxml==4.9.3
Markdown==3.4.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
msgpack==1.0.5
nest-asyncio==1.5.6
numpy==1.25.2
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.0
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.36
proto-plus==1.22.2
protobuf==4.23.3
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.14.0
PyJWT==2.7.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.6
pywin32==305
pyzmq==25.0.0
redis==4.5.5
reportlab==4.0.4
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
seaborn==0.12.2
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.3
stack-data==0.6.2
svglib==1.5.1
telepath==0.3.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.8.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.7.1
tzdata==2022.6
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.16.6
virtualenvwrapper-win==1.2.7
wagtail==5.0.2
wcwidth==0.2.6
webdriver-manager==3.8.5
webencodings==0.5.1
Willow==1.5.1
wsproto==1.2.0
